--- a/cmd/chihaya/main.go
+++ b/cmd/chihaya/main.go
@@ -56,11 +56,11 @@ func (r *Run) Start(ps storage.PeerStore) error {
 	r.sg.Add(prometheus.NewServer(cfg.PrometheusAddr))
 
 	if ps == nil {
-		log.WithFields(cfg.Storage.LogFields()).Info("starting storage")
 		ps, err = storage.NewPeerStore(cfg.Storage.Name, cfg.Storage.Config)
 		if err != nil {
 			return errors.New("failed to create memory storage: " + err.Error())
 		}
+		log.WithFields(ps.LogFields()).Info("started storage")
 	}
 	r.peerStore = ps