This is a WIP branch attempting to fix #267.