@mrd0ll4r 2017-02-22
Oh I forgot: Also needs docs for the config and something that describes how peers are selected