@jzelinskie 2017-05-04
Rebased and addressed a bunch of comments and ran benchmarks:

```
BenchmarkPut-2                                   3000000               465 ns/op
BenchmarkPut1k-2                                 3000000               553 ns/op
BenchmarkPut1kInfohash-2                         5000000               386 ns/op
BenchmarkPut1kInfohash1k-2                       5000000               362 ns/op
BenchmarkPutDelete-2                             1000000              1552 ns/op
BenchmarkPutDelete1k-2                           1000000              1571 ns/op
BenchmarkPutDelete1kInfohash-2                   1000000              1682 ns/op
BenchmarkPutDelete1kInfohash1k-2                 1000000              1672 ns/op
BenchmarkDeleteNonexist-2                       10000000               120 ns/op
BenchmarkDeleteNonexist1k-2                     10000000               113 ns/op
BenchmarkDeleteNonexist1kInfohash-2             20000000                96.8 ns/op
BenchmarkDeleteNonexist1kInfohash1k-2           20000000               100 ns/op
BenchmarkPutGradDelete-2                          500000              2516 ns/op
BenchmarkPutGradDelete1k-2                        500000              2683 ns/op
BenchmarkPutGradDelete1kInfohash-2                500000              2685 ns/op
BenchmarkPutGradDelete1kInfohash1k-2              500000              2866 ns/op
BenchmarkGradNonexist-2                          3000000               510 ns/op
BenchmarkGradNonexist1k-2                        2000000               583 ns/op
BenchmarkGradNonexist1kInfohash-2                5000000               365 ns/op
BenchmarkGradNonexist1kInfohash1k-2              5000000               384 ns/op
BenchmarkAnnounceLeecher-2                         50000             44568 ns/op
BenchmarkAnnounceLeecher1kInfohash-2               20000             86686 ns/op
BenchmarkAnnounceSeeder-2                          50000             40797 ns/op
BenchmarkAnnounceSeeder1kInfohash-2                20000             91576 ns/op
PASS
ok      github.com/chihaya/chihaya/storage/memorybysubnet       68.414s

BenchmarkPut-2                                  10000000               232 ns/op
BenchmarkPut1k-2                                 5000000               256 ns/op
BenchmarkPut1kInfohash-2                        10000000               177 ns/op
BenchmarkPut1kInfohash1k-2                      10000000               182 ns/op
BenchmarkPutDelete-2                             2000000               816 ns/op
BenchmarkPutDelete1k-2                           2000000               754 ns/op
BenchmarkPutDelete1kInfohash-2                   2000000               802 ns/op
BenchmarkPutDelete1kInfohash1k-2                 2000000               794 ns/op
BenchmarkDeleteNonexist-2                       10000000               119 ns/op
BenchmarkDeleteNonexist1k-2                     10000000               112 ns/op
BenchmarkDeleteNonexist1kInfohash-2             20000000                86.6 ns/op
BenchmarkDeleteNonexist1kInfohash1k-2           20000000                92.5 ns/op
BenchmarkPutGradDelete-2                         1000000              1167 ns/op
BenchmarkPutGradDelete1k-2                       1000000              1167 ns/op
BenchmarkPutGradDelete1kInfohash-2               1000000              1237 ns/op
BenchmarkPutGradDelete1kInfohash1k-2             1000000              1232 ns/op
BenchmarkGradNonexist-2                          5000000               261 ns/op
BenchmarkGradNonexist1k-2                        5000000               269 ns/op
BenchmarkGradNonexist1kInfohash-2               10000000               200 ns/op
BenchmarkGradNonexist1kInfohash1k-2             10000000               197 ns/op
BenchmarkAnnounceLeecher-2                        200000              6028 ns/op
BenchmarkAnnounceLeecher1kInfohash-2              200000              8268 ns/op
BenchmarkAnnounceSeeder-2                         200000              6095 ns/op
BenchmarkAnnounceSeeder1kInfohash-2               200000              7750 ns/op
PASS
ok      github.com/chihaya/chihaya/storage/memory       49.556s
```