@mrd0ll4r 2017-05-04
Benchcmp:

```
benchmark                                 old ns/op     new ns/op     delta
BenchmarkPut-2                            232           465           +100.43%
BenchmarkPut1k-2                          256           553           +116.02%
BenchmarkPut1kInfohash-2                  177           386           +118.08%
BenchmarkPut1kInfohash1k-2                182           362           +98.90%
BenchmarkPutDelete-2                      816           1552          +90.20%
BenchmarkPutDelete1k-2                    754           1571          +108.36%
BenchmarkPutDelete1kInfohash-2            802           1682          +109.73%
BenchmarkPutDelete1kInfohash1k-2          794           1672          +110.58%
BenchmarkDeleteNonexist-2                 119           120           +0.84%
BenchmarkDeleteNonexist1k-2               112           113           +0.89%
BenchmarkDeleteNonexist1kInfohash-2       86.6          96.8          +11.78%
BenchmarkDeleteNonexist1kInfohash1k-2     92.5          100           +8.11%
BenchmarkPutGradDelete-2                  1167          2516          +115.60%
BenchmarkPutGradDelete1k-2                1167          2683          +129.91%
BenchmarkPutGradDelete1kInfohash-2        1237          2685          +117.06%
BenchmarkPutGradDelete1kInfohash1k-2      1232          2866          +132.63%
BenchmarkGradNonexist-2                   261           510           +95.40%
BenchmarkGradNonexist1k-2                 269           583           +116.73%
BenchmarkGradNonexist1kInfohash-2         200           365           +82.50%
BenchmarkGradNonexist1kInfohash1k-2       197           384           +94.92%
BenchmarkAnnounceLeecher-2                6028          44568         +639.35%
BenchmarkAnnounceLeecher1kInfohash-2      8268          86686         +948.45%
BenchmarkAnnounceSeeder-2                 6095          40797         +569.35%
BenchmarkAnnounceSeeder1kInfohash-2       7750          91576         +1081.63%
```